[client] show battle info on hover Fixes #1865
